+++ core/src/main/java/nl/nn/adapterframework/http/RestServiceDispatcher.java
@@ -21,6 +21,8 @@ import java.util.Map;
 import java.util.SortedMap;
 import java.util.TreeMap;
 
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import nl.nn.adapterframework.configuration.ConfigurationException;
@@ -65,8 +67,8 @@ public class RestServiceDispatcher  {
 	 * @param request the <code>String</code> with the request/input
 	 * @return String with the result of processing the <code>request</code> throught the <code>serviceName</code>
      */
-	public String dispatchRequest(String restPath, String uri, String method, String etag, String contentType, String request, Map context, HttpServletResponse httpServletResponse) throws ListenerException {
-		
+	public String dispatchRequest(String restPath, String uri, HttpServletRequest httpServletRequest, String etag, String contentType, String request, Map context, HttpServletResponse httpServletResponse, ServletContext servletContext) throws ListenerException {
+		String method = httpServletRequest.getMethod();
 		if (log.isDebugEnabled()) log.debug("searching listener for uri ["+uri+"] method ["+method+"]");
 		
 		String matchingPattern=null;
@@ -105,7 +107,9 @@ public class RestServiceDispatcher  {
 		if (etagKey!=null) context.put(etagKey,etag);
 		if (contentTypeKey!=null) context.put(contentTypeKey,contentType);
 		if (log.isDebugEnabled()) log.debug("dispatching request, uri ["+uri+"] listener pattern ["+matchingPattern+"] method ["+method+"] etag ["+etag+"] contentType ["+contentType+"]");
+		if (httpServletRequest!=null) context.put("restListenerServletRequest", httpServletRequest);
 		if (httpServletResponse!=null) context.put("restListenerServletResponse", httpServletResponse);
+		if (servletContext!=null) context.put("restListenerServletContext", servletContext);
 
 		if (listener instanceof RestListener) {
 			RestListener restListener = (RestListener) listener;